Now with 4 options: timed waves, personal timer, side tickets, personal tickets (ticket types stack so can have personal pool and then side-based 'overflow' pool). Respawn timer is variable-based so can be altered mid-mission.
