+++ components/respawn/functions.hpp
@@ -1,4 +1,5 @@
 class respawn
 {
     file = "components\respawn";
+    class getPlayerRespawnTime{};
 };
