+++ components/respawn/loadout/respawnTemplates.hpp
@@ -1,0 +1,5 @@
+class CAFE_Loadout
+{
+	onPlayerRespawn = "components\respawn\loadout\onPlayerRespawn.sqf";
+	isCall = 0;
+};
