+++ components/respawn/squad/respawnTemplates.hpp
@@ -1,0 +1,6 @@
+class CAFE_Squad
+{
+	onPlayerKilled = "components\respawn\squad\onPlayerKilled.sqf";
+	onPlayerRespawn = "components\respawn\squad\onPlayerRespawn.sqf";
+	isCall = 1;
+};
