+++ components/respawn/timed/macros.hpp
@@ -1,0 +1,1 @@
+#include "../macros.hpp"
