+++ components/respawn/timed/respawnTemplates.hpp
@@ -1,0 +1,5 @@
+class CAFE_Timed
+{
+	onPlayerKilled = "components\respawn\timed\onPlayerKilledOrRespawned.sqf";
+	onPlayerRespawn = "components\respawn\timed\onPlayerKilledOrRespawned.sqf";
+};
