+++ components/respawn/wave/respawnTemplates.hpp
@@ -1,0 +1,6 @@
+class CAFE_Wave
+{
+	onPlayerKilled = "components\respawn\wave\onPlayerKilledOrRespawned.sqf";
+	onPlayerRespawn = "components\respawn\wave\onPlayerKilledOrRespawned.sqf";
+	respawnTypes[] = {2,3};
+};
