+++ configuration/respawn.hpp
@@ -13,19 +13,22 @@
 // When a player respawns, do they get to teleport to a group of their choosing?
 // #define ALLOW_TELEPORT_UPON_RESPAWN
 
+#define RESPAWN_MODE_BLUFOR         	    RESPAWN_MODE_TIMED
+#define RESPAWN_DELAY_BLUFOR        	    30
+#define RESPAWN_SIDE_TICKETS_BLUFOR		    30
+#define RESPAWN_PLAYER_TICKETS_BLUFOR	    2
+       
+#define RESPAWN_MODE_OPFOR          	    {RESPAWN_DEFAULTS, RESPAWN_TIMED_WAVE}
+#define RESPAWN_DELAY_OPFOR         	    10
+#define RESPAWN_SIDE_TICKETS_OPFOR		    30
+#define RESPAWN_PLAYER_TICKETS_OPFOR	    2
+    
+#define RESPAWN_MODE_INDFOR         	    {RESPAWN_DEFAULTS, RESPAWN_TIMED_WAVE}
+#define RESPAWN_DELAY_INDFOR        	    RESPAWN_DELAY_BLUFOR
+#define RESPAWN_SIDE_TICKETS_INDFOR		    30
+#define RESPAWN_PLAYER_TICKETS_INDFOR	    2
 
-#define RESPAWN_MODES_BLUFOR        {RESPAWN_DEFAULTS(BLUFOR), RESPAWN_WAVE, RESPAWN_TICKETS}
-#define RESPAWN_DELAY_BLUFOR        30
-#define RESPAWN_AT_START_BLUFOR     1
-
-#define RESPAWN_MODES_OPFOR         {RESPAWN_DEFAULTS(OPFOR), RESPAWN_WAVE}
-#define RESPAWN_DELAY_OPFOR         10
-#define RESPAWN_AT_START_OPFOR      RESPAWN_AT_START_BLUFOR
-
-#define RESPAWN_MODES_INDFOR        {RESPAWN_DEFAULTS(INDFOR), RESPAWN_WAVE}
-#define RESPAWN_DELAY_INDFOR        RESPAWN_DELAY_BLUFOR
-#define RESPAWN_AT_START_INDFOR     RESPAWN_AT_START_BLUFOR
-
-#define RESPAWN_MODES_CIVILIAN      {RESPAWN_DEFAULTS(CIVILIAN)}
-#define RESPAWN_DELAY_CIVILIAN      5
-#define RESPAWN_AT_START_CIVILIAN   RESPAWN_AT_START_BLUFOR
+#define RESPAWN_MODE_CIVILIAN       		{RESPAWN_DEFAULTS}
+#define RESPAWN_DELAY_CIVILIAN      		5
+#define RESPAWN_SIDE_TICKETS_CIVILIAN		30
+#define RESPAWN_PLAYER_TICKETS_CIVILIAN		2
