+++ respawn_macros.hpp
@@ -1,12 +1,12 @@
-#define RESPAWN_WAVE "Wave"
+#define RESPAWN_TIMED_WAVE "CAFE_Wave"
 #define RESPAWN_TICKETS "Tickets"
+#define RESPAWN_TIMED "CAFE_Timed"
 
-#define RESPAWN_MODE_TIMED       {"MenuPosition", "Spectator", "CAFE_Loadout", "CAFE_Squad"}
-#define RESPAWN_MODE_TIMED_WAVES {"MenuPosition", "Spectator", "Wave", "CAFE_Loadout", "CAFE_Squad"}
-#define RESPAWN_MODE_TIMED_WAVES {"MenuPosition", "Spectator", "Wave", "CAFE_Loadout", "CAFE_Squad"}
+#define RESPAWN_MODE_TIMED       			{"MenuPosition", "Spectator", RESPAWN_TIMED, "CAFE_Loadout", "CAFE_Squad"}
+#define RESPAWN_MODE_TIMED_WAVES 			{"MenuPosition", "Spectator", RESPAWN_TIMED_WAVE, "CAFE_Loadout", "CAFE_Squad"}
+#define RESPAWN_MODE_TICKETS 				{"MenuPosition", "Spectator", RESPAWN_TICKETS, "CAFE_Loadout", "CAFE_Squad"}
+#define RESPAWN_MODE_TIMED_WAVES_TICKETS 	{"MenuPosition", "Spectator", RESPAWN_TIMED_WAVE, RESPAWN_TICKETS, "CAFE_Loadout", "CAFE_Squad"}
 
-#define RESPAWN_DEFAULTS(SIDE) STRING(CONCAT(CAFE_RespawnConfig_,SIDE)), "MenuPosition", "Spectator", "CAFE_Loadout", "CAFE_Squad"
-
-#define RESPAWN_MODES(SIDE) CONCAT(RESPAWN_ARRAY_,SIDE)
+#define RESPAWN_DEFAULTS "MenuPosition", "Spectator", "CAFE_Loadout", "CAFE_Squad"
 
 #define DEFAULT_RESPAWN_ARRAY {RESPAWN_DEFAULTS}
